+++ BMLTiOSLib/Test Harness Project/BMLTiOSTesterMarker.swift
@@ -81,7 +81,7 @@ class BMLTiOSLibTesterAnnotation: NSObject, MKAnnotation, NSCoding {
      */
     @objc required init?(coder aDecoder: NSCoder) {
         self.meetings = (aDecoder.decodeObject(forKey: self.sMeetingsObjectKey) as? BMLTiOSLibTesterMeetingList)!
-        if let tempCoordinate = aDecoder.decodeObject(forKey: self.sCoordinateObjectKey) as? [NSNumber]! {
+        if let tempCoordinate = aDecoder.decodeObject(forKey: self.sCoordinateObjectKey) as? [NSNumber] {
             self.coordinate.longitude = tempCoordinate[0].doubleValue
             self.coordinate.latitude = tempCoordinate[1].doubleValue
         }